# improved topological sort based on Kahn's algorithm