Benchmarking Topological Sort Implementations
---------------------------------------------

Graph size: 10, Edge probability: 0.1
Old implementation: 0.009 ms
Optimized implementation: 0.006 ms
Improvement: 29.71%

Graph size: 10, Edge probability: 0.3
Old implementation: 0.004 ms
Optimized implementation: 0.002 ms
Improvement: 42.62%

Graph size: 10, Edge probability: 0.5
Old implementation: 0.003 ms
Optimized implementation: 0.001 ms
Improvement: 57.82%

Graph size: 100, Edge probability: 0.1
Old implementation: 0.033 ms
Optimized implementation: 0.024 ms
Improvement: 25.64%

Graph size: 100, Edge probability: 0.3
Old implementation: 0.035 ms
Optimized implementation: 0.021 ms
Improvement: 39.79%

Graph size: 100, Edge probability: 0.5
Old implementation: 0.034 ms
Optimized implementation: 0.020 ms
Improvement: 41.76%

Optimized implementation: 0.366 ms

Graph size: 1000, Edge probability: 0.3
Old implementation: 0.515 ms
Optimized implementation: 0.325 ms
Improvement: 36.96%

Graph size: 1000, Edge probability: 0.5
Optimized implementation: 0.338 ms
Improvement: 29.18%

Graph size: 10000, Edge probability: 0.1
Old implementation: 14.470 ms
Optimized implementation: 4.523 ms
Improvement: 68.74%

Graph size: 10000, Edge probability: 0.3
Old implementation: 14.812 ms
Optimized implementation: 4.812 ms
Improvement: 67.51%

Graph size: 10000, Edge probability: 0.5
Old implementation: 14.726 ms
Optimized implementation: 4.940 ms
Improvement: 66.45%